+++ sw_projects/P2_app/protocol2_control.h
@@ -19,6 +19,7 @@ typedef struct
 TP2RunState P2DecodeRunState(uint8_t Flags);
 void P2SequenceReset(TP2SequenceTracker *Tracker);
 bool P2SequenceAccept(TP2SequenceTracker *Tracker, uint32_t Sequence, uint32_t *MissingPackets);
+bool P2ControlSequenceAccept(TP2SequenceTracker *Tracker, uint32_t Sequence, uint32_t *MissingPackets);
 uint16_t P2ScaleFifoSamples(uint32_t Locations, uint32_t SamplesPerGroup, uint32_t LocationsPerGroup);
 
 #endif